requirements.txt
pandas==1.4.4
python_binance==1.0.17
streamlit==1.29.0
